Updated behave tests for power event subscribers
